+++ docs/releases.md
@@ -1,5 +1,10 @@
 # Releases
 
+## TODO
+* [Published](https://pypi.org/project/itsh5py/TODO) on TODO
+* Added support for the new config `max_tree_children` to reduce too large tree
+  views with a default of 30 elements per group.
+
 ## 0.7.1
 * [Published](https://pypi.org/project/itsh5py/0.7.1) on 2022-2-2
 * Changed affiliations and contact for main author